Add NVIDIA Blackwell (RTX 5080) GPU support for TRELLIS
Migrate TRELLIS 3D generation pipeline to support Blackwell architecture
(compute 12.0, sm_120). Key changes:

- Add SDPA attention backend (xformers crashes on Blackwell)
- Add VRAMManager with dynamic dtype and model swapping for 16GB VRAM
- Fix float16 dtype mismatches throughout the pipeline
- Add text-to-3D app with variant editor UI
- Rebuild all CUDA extensions with sm_120 support
- Document 16 errors and fixes in ERRORS_AND_FIXES.md

Tested on RTX 5080 (16GB), Ubuntu 24.04, CUDA 12.8, PyTorch 2.7.0+cu128.

+++ trellis/renderers/mesh_renderer.py
@@ -91,7 +91,7 @@ class MeshRenderer:
         RT = extrinsics.unsqueeze(0)
         full_proj = (perspective @ extrinsics).unsqueeze(0)
         
-        vertices = mesh.vertices.unsqueeze(0)
+        vertices = mesh.vertices.float().unsqueeze(0)
 
         vertices_homo = torch.cat([vertices, torch.ones_like(vertices[..., :1])], dim=-1)
         vertices_camera = torch.bmm(vertices_homo, RT.transpose(-1, -2))
